% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{antimalarialQualitySurvey}
\alias{antimalarialQualitySurvey}
\title{Antimalarial Quality Survey Data}
\format{A data frame with 9 rows and 7 variables}
\source{
\itemize{
  \item {
     \href{http://apps.who.int/medicinedocs/documents/s17069e/s17069e.pdf}{US Pharmacopeia. Survey of the Quality of Selected Antimalarial Medicines Circulating in Madagascar, Senegal, and Uganda, 2009}
  }
  \item {
     \href{http://www.who.int/medicines/publications/WHO_QAMSA_report.pdf}{World Health Organization. Survey of the Quality of Selected Antimalarial Medicines Circulating in Six Countries of Sub-Saharan Africa, 2011}
  }
}
}
\description{
\itemize{
  \item Country.
  \item Failed. Number of samples that failed quality testing.
  \item Tested. Number of tested samples.
  \item Mean. Proportion of tested samples that failed quality testing.
  \item Var.
  \item SE.
  \item Order. Index number used by webtool().
	}
}

