% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{pqantimalarials}
\alias{pqantimalarials}
\alias{pqantimalarials-package}
\title{Estimated under-five deaths associated with poor-quality
  antimalarials in sub-Saharan Africa.}
\description{
pqantimalarials provides an interactive web tool that
allows users to produce their own estimates of the
number of under-five deaths associated with poor-quality
antimalarials in sub-Saharan Africa according
to the methodology of Renschler et al. (2014). Users
can set the input paramaters and download the results
from uncertainty and sensitivity analyses.
The \code{webtool()} function
starts the interactive web tool on the user's
machine. You can follow the development of this package at:
\url{https://github.com/renschler/pqantimalarials}.
}
\references{
J. Patrick Renschler, Kelsey Walters, Paul Newton,
  Ramanan Laxminarayan. "Estimated under-five deaths
  associated with poor-quality antimalarials in
  sub-Sarahan Africa". 2014.
}

