% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{privateSales}
\alias{privateSales}
\title{Estimated 2013 private sector antimalarial sales to under-five malaria
  positive children.}
\format{A data frame with 39 rows and 3 variables}
\source{
J. Cohen, A. Woolsey, O. Sabot, P. Gething, A. Tatem, B. Moonen.
  "Optimizing Investments in Malaria Treatment and Diagnosis."
  Science 338 (2012): 612-614.
}
\description{
\itemize{
  \item Country.
  \item u5Sales. Median estimated 2013 private sector antimalarial
    sales to under-five P. falciparum  positive children (Cohen et al. 2012).
  \item StDev. Standard deviation of estimated 2013 private
     sector antimalarial sales to under-five P. falciparum positive
     children. Standard deviation was calculated using the
     interquartile range provided by Cohen et al. (2012), assuming
     a normal distribution.
}
}

