% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{u5deathsWHO2012}
\alias{u5deathsWHO2012}
\title{World Health Organization 2012 under-five mortality estimates.}
\format{A data frame with 39 rows and 4 variables.}
\source{
\href{http://apps.who.int/gho/data/node.main.525}{MDG 4: Child health: Under-five mortality by country}
}
\description{
\itemize{
  \item Country.
  \item Year.
  \item Deaths. WHO estimated number of under-five deaths (thousands).
  \item MortalityRate. WHO estimated probability of dying by age 5
    per 1000 live births.
}
}

