% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{u5malariadeathsWHO2010}
\alias{u5malariadeathsWHO2010}
\title{World Health Organization estimated 2010 under-five malaria deaths.}
\format{A data frame with 193 rows and 3 variables.}
\source{
\href{http://apps.who.int/gho/data/node.main.GBDC-YEARS0-4?lang=en}{WHO - Global Health Observatory Data Repository: Mortality By age group and cause: Children aged 0 to 4 years 2010}
}
\description{
\itemize{
  \item Year.
  \item Country.
  \item Numeric. WHO estimated 2010 under-five malaria deaths.
}
}

